changed E2G start and end types, updated target prioritisation script

+++ config/clickhouse/schema/enhancer_to_gene.sql
@@ -26,8 +26,8 @@ CREATE TABLE if not exists enhancer_to_gene_log (
         'Y',
         'MT'
     ),
-    `start` Int64,
-    `end` Int64,
+    `start` UInt32,
+    `end` UInt32,
     `geneId` String,
     `biosampleName` LowCardinality (String),
     `biosampleId` LowCardinality (String),
